requirements.txt
requests
discord.py
bravado
